TestDriver.java
A draft foundation for the test driver (secondary driver) class. I left some methods to be created/modified. I made sure to add a bunch of comments just to make it easy to understand for development purposes, and then we will clean the comments later, once done with the class itself.